Shader Loader
After looking at the tutorial I decided to take a crack at making my own shader loader and not just using the tutorials code. After a while of setting it up it seems like the vertex nor the fragment was working. Embarrassingly this is because I somehow had the same code for the fragment and vertex. After adding the vertex code to its proper file I was able to get the shaders working again.

Side note: Added a gitignore