Reorganize + Textures!!!
Followed the tutorial all the way up to the wooden container texture tutorial. In addition I've changed the folder layout to be more organized. Ive decided to use stb_image for ease of use. Personal includes are now in a include folder inside the project. Includes such as GLFW and Glad are not included and you will need to download and include them yourselves.

Note: The path variables are set up for windows so if you are on linux you may need to change the path to fit your OS.